Adjust init order
