+++ jort/__init__.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import, division, print_function
 from ._version import __version__
 
 from .exceptions import JortException, JortCredentialException
-from .config import init
+from .config import init, config_general, config_email, config_text
 from .tracker import Tracker, track
 from .track_cli import track_new, track_existing
 from .reporting_callbacks import EmailNotification, TextNotification, PrintReport
